backend/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0

# Google Cloud services
google-cloud-speech>=2.20.0
google-cloud-texttospeech>=2.14.0
google-generativeai>=0.3.0

# RAG support
sentence-transformers>=2.2.0
qdrant-client>=1.6.0

# Example/launch script dependencies
websockets>=12.0
sounddevice>=0.4.6
httpx>=0.25.0
numpy>=1.24.0

# Testing
pytest>=7.4.0

